Create Code 1
An array stores the details of customer transactions (account number, name of account holder, transaction type and amount) in a bank. Transaction type would be either Withdrawal (W or w) and Deposit (D or d). Use suitable data types for array creation.

Write a Java program to create an array and store the details of 5 customer transactions. Using the stored details, calculate the following:

                           i.          Total deposit amount

                          ii.          Total withdrawal amount

                         iii.          Count of all deposit transactions

                         iv.          Count of all withdrawal transactions

                         v.          Name of customer who has deposited the 
                                      maximum amount

                       vi.          Name of the customer who has withdrawn 
                                     the minimum amount